more splitting of rendering code
