+++ x-pack/platform/plugins/private/data_federation/public/create_data_source_flyout/data_sources/gcs/create_data_source_flyout_type_settings_gcs.tsx
@@ -18,7 +18,7 @@ import {
 
 import type { UseFormUnregister } from 'react-hook-form';
 import { type Control, useController } from 'react-hook-form';
-import type { CreateDataSourceFlyoutFormValues } from './types';
+import type { CreateDataSourceFlyoutFormValues } from '../../types';
 
 export function CreateDataSourceFlyoutTypeSettingsGcs({
   control,
