+++ x-pack/platform/plugins/private/data_federation/public/create_data_source_flyout/data_sources/s3/s3_data_source_advanced_settings.tsx
@@ -1,0 +1,50 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import React, { useEffect } from 'react';
+import { i18n } from '@kbn/i18n';
+import { EuiFieldText, EuiFormRow } from '@elastic/eui';
+import { type Control, type UseFormUnregister, useController } from 'react-hook-form';
+import type { CreateDataSourceFlyoutFormValues } from '../../types';
+
+export function S3DataSourceAdvancedSettings({
+  control,
+  unregister,
+}: {
+  control: Control<CreateDataSourceFlyoutFormValues>;
+  unregister: UseFormUnregister<CreateDataSourceFlyoutFormValues>;
+}) {
+  const { field: endpointField } = useController({
+    name: 'settings.endpoint',
+    control,
+  });
+
+  useEffect(() => {
+    return () => {
+      unregister('settings.endpoint');
+    };
+  }, [unregister]);
+
+  return (
+    <EuiFormRow
+      label={i18n.translate('xpack.dataFederation.createFlyout.s3.fields.endpoint', {
+        defaultMessage: 'Endpoint',
+      })}
+      fullWidth
+    >
+      <EuiFieldText
+        data-test-subj="createDataSourceFlyoutS3Endpoint"
+        fullWidth
+        autoComplete="off"
+        value={endpointField.value}
+        onChange={(e) => endpointField.onChange(e.target.value)}
+        name={endpointField.name}
+        inputRef={endpointField.ref}
+      />
+    </EuiFormRow>
+  );
+}
