+++ x-pack/platform/plugins/private/data_federation/public/create_data_source_flyout/data_sources/s3/s3_data_source_required_settings.tsx
@@ -21,7 +21,7 @@ import type { CreateDataSourceFlyoutFormValues } from '../../types';
 import type { FederatedIdentityClusterInfo } from './federated_identity_cluster_info';
 import { FederatedIdentityClusterInfoFields } from './federated_identity_cluster_info_fields';
 
-export function CreateDataSourceFlyoutTypeSettingsS3Region({
+export function S3DataSourceRequiredSettings({
   control,
   unregister,
   isRequired,
@@ -69,44 +69,6 @@ export function CreateDataSourceFlyoutTypeSettingsS3Region({
         onChange={(e) => regionField.onChange(e.target.value)}
         name={regionField.name}
         inputRef={regionField.ref}
-      />
-    </EuiFormRow>
-  );
-}
-
-export function CreateDataSourceFlyoutTypeSettingsS3({
-  control,
-  unregister,
-}: {
-  control: Control<CreateDataSourceFlyoutFormValues>;
-  unregister: UseFormUnregister<CreateDataSourceFlyoutFormValues>;
-}) {
-  const { field: endpointField } = useController({
-    name: 'settings.endpoint',
-    control,
-  });
-
-  useEffect(() => {
-    return () => {
-      unregister('settings.endpoint');
-    };
-  }, [unregister]);
-
-  return (
-    <EuiFormRow
-      label={i18n.translate('xpack.dataFederation.createFlyout.s3.fields.endpoint', {
-        defaultMessage: 'Endpoint',
-      })}
-      fullWidth
-    >
-      <EuiFieldText
-        data-test-subj="createDataSourceFlyoutS3Endpoint"
-        fullWidth
-        autoComplete="off"
-        value={endpointField.value}
-        onChange={(e) => endpointField.onChange(e.target.value)}
-        name={endpointField.name}
-        inputRef={endpointField.ref}
       />
     </EuiFormRow>
   );
